requirements.prod.txt
# Production Requirements
# Install with: pip install -r requirements.prod.txt

# Core dependencies from requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlmodel==0.0.24
psycopg[pool]==3.2.9
asyncpg==0.30.0
twilio==9.7.0
python-dotenv==1.1.1
requests==2.32.4
httpx==0.28.1
jinja2==3.1.6
python-multipart==0.0.20

# Production-specific additions
gunicorn==21.2.0
supervisor==4.2.5

# Security enhancements
python-jose[cryptography]==3.3.0

# Monitoring (optional but recommended)
python-multipart==0.0.20
